+++ ThirdPerson/blueprints/blu_ground.ts
@@ -1,6 +1,6 @@
 import {float} from "../../common/random.js";
 import {collide} from "../components/com_collide.js";
-import {render_diffuse} from "../components/com_render_diffuse.js";
+import {render_diffuse} from "../components/com_render1.js";
 import {rigid_body} from "../components/com_rigid_body.js";
 import {Blueprint} from "../core.js";
 import {Game, Layer} from "../game.js";
