+++ ThirdPerson/game.ts
@@ -13,7 +13,7 @@ import {sys_light} from "./systems/sys_light.js";
 import {sys_mimic} from "./systems/sys_mimic.js";
 import {sys_move} from "./systems/sys_move.js";
 import {sys_physics} from "./systems/sys_physics.js";
-import {sys_render} from "./systems/sys_render.js";
+import {sys_render} from "./systems/sys_render1.js";
 import {sys_transform} from "./systems/sys_transform.js";
 import {World} from "./world.js";
